+++ frontend/app_flowy/lib/plugins/grid/presentation/widgets/filter/filter_info.dart
@@ -2,21 +2,22 @@ import 'package:app_flowy/plugins/grid/application/field/field_controller.dart';
 import 'package:flowy_sdk/protobuf/flowy-grid/checkbox_filter.pb.dart';
 import 'package:flowy_sdk/protobuf/flowy-grid/date_filter.pb.dart';
 import 'package:flowy_sdk/protobuf/flowy-grid/field_entities.pb.dart';
+import 'package:flowy_sdk/protobuf/flowy-grid/select_option_filter.pbserver.dart';
 import 'package:flowy_sdk/protobuf/flowy-grid/text_filter.pb.dart';
 import 'package:flowy_sdk/protobuf/flowy-grid/util.pb.dart';
 
 class FilterInfo {
   final String viewId;
   final FilterPB filter;
-  final FieldInfo field;
+  final FieldInfo fieldInfo;
 
-  FilterInfo(this.viewId, this.filter, this.field);
+  FilterInfo(this.viewId, this.filter, this.fieldInfo);
 
-  FilterInfo copyWith({FilterPB? filter, FieldInfo? field}) {
+  FilterInfo copyWith({FilterPB? filter, FieldInfo? fieldInfo}) {
     return FilterInfo(
       viewId,
       filter ?? this.filter,
-      field ?? this.field,
+      fieldInfo ?? this.fieldInfo,
     );
   }
 
@@ -39,5 +40,14 @@ class FilterInfo {
       return null;
     }
     return CheckboxFilterPB.fromBuffer(filter.data);
+  }
+
+  SelectOptionFilterPB? selectOptionFilter() {
+    if (filter.fieldType == FieldType.SingleSelect ||
+        filter.fieldType == FieldType.MultiSelect) {
+      return SelectOptionFilterPB.fromBuffer(filter.data);
+    } else {
+      return null;
+    }
   }
 }