+++ frontend/app_flowy/lib/plugins/grid/presentation/widgets/filter/menu_item.dart
@@ -4,7 +4,7 @@ import 'package:flutter/material.dart';
 import 'choicechip/checkbox.dart';
 import 'choicechip/date.dart';
 import 'choicechip/number.dart';
-import 'choicechip/select_option.dart';
+import 'choicechip/select_option/select_option.dart';
 import 'choicechip/text.dart';
 import 'choicechip/url.dart';
 import 'filter_info.dart';
@@ -20,7 +20,7 @@ class FilterMenuItem extends StatelessWidget {
 }
 
 Widget buildFilterChoicechip(FilterInfo filterInfo) {
-  switch (filterInfo.field.fieldType) {
+  switch (filterInfo.fieldInfo.fieldType) {
     case FieldType.Checkbox:
       return CheckboxFilterChoicechip(filterInfo: filterInfo);
     case FieldType.DateTime: